@@ -1,216 +1,261 @@
-require 'spec_helper'
+require 'rails_helper'
 
-# This spec was generated by rspec-rails when you ran the scaffold generator.
-# It demonstrates how one might use RSpec to specify the controller code that
-# was generated by Rails when you ran the scaffold generator.
-#
-# It assumes that the implementation code is generated by the rails scaffold
-# generator.  If you are using any extension libraries to generate different
-# controller code, this generated spec may or may not pass.
-#
-# It only uses APIs available in rails and/or rspec-rails.  There are a number
-# of tools you can use to make these specs even more expressive, but we're
-# sticking to rails and rspec-rails APIs to keep things simple and stable.
-#
-# Compared to earlier versions of this generator, there is very limited use of
-# stubs and message expectations in this spec.  Stubs are only used when there
-# is no simpler way to get a handle on the object needed for the example.
-# Message expectations are only used when there is no simpler way to specify
-# that an instance is receiving a specific message.
+describe PagesController, :type => :controller do
+  let(:page) { mock_model Page, id: '2' }
 
-describe PagesController do
+  before { allow(controller).to receive(:superadmin?) { true } }
 
-  # This should return the minimal set of attributes required to create a valid
-  # Page. As you add validations to Page, be sure to
-  # adjust the attributes here as well.
-  let(:valid_attributes) { { "name" => "MyString", "permalink" => "about" } }
-
-  # This should return the minimal set of values that should be in the session
-  # in order to pass any filters (e.g. authentication) defined in
-  # PagesController. Be sure to keep this updated too.
-  let(:valid_session) { {} }
-
-  before(:each) do
-    controller.stub(:admin?).and_return(true)
-  end
-
-  describe "GET index" do
-    it "assigns all pages as @pages" do
-      page = Page.create! valid_attributes
+  describe 'GET index' do
+    it 'is restricted' do
+      expect(controller).to receive(:superadmin?) { false }
       get :index, {}
-      assigns(:pages).should eq([page])
+      expect(response.status).to eq 302
     end
 
-    it 'should use a full-width layout' do
+    it 'uses a full-width layout' do
       get :index, {}
-      response.should render_template 'layouts/full_width'
+      expect(response).to render_template 'layouts/full_width'
+    end
+
+    it "assigns the pages in alphabetical order by default" do
+      pages = double Array
+      expect(Page).to receive(:order).with('name ASC').and_return pages
+      get :index, {}
+      expect(assigns(:pages)).to eq(pages)
     end
   end
 
-  describe "GET show" do
-    before :each do
-      admin_user = double('User')
-      admin_user.stub(:admin?).and_return(true)
-      controller.stub(:current_user).and_return(admin_user)
-      @valid_page = double('Page', :name => 'About Us', :permalink => 'about', :content => 'blah blah')
-      @error_page = double('Page')
+  describe 'GET show' do
+    let(:about_page) { double :page,
+                              name: 'About Us',
+                              permalink: 'about',
+                              content: 'blah blah'
+    }
+    let(:user) { double :user }
+
+    describe 'assigns site superadmin status of current_user to @superadmin' do
+      it 'NIL when there is no current_user' do
+        get :show, {id: 'about'}
+        expect(assigns(:superadmin)).to be nil
+      end
+      it 'FALSE when current_user is NOT superadmin' do
+        allow(user).to receive(:superadmin?) { false }
+        allow(controller).to receive(:current_user) { user }
+        get :show, {id: 'about'}
+        expect(assigns(:superadmin)).to be false
+      end
+
+      it 'TRUE when current_user is superadmin' do
+        allow(user).to receive(:superadmin?) { true }
+        allow(controller).to receive(:current_user) { user }
+        get :show, {id: 'about'}
+        expect(assigns(:superadmin)).to be true
+      end
     end
 
-    after(:each) {response.should render_template 'layouts/full_width'}
-
-    it 'assigns @admin if current_user is admin' do
-      Page.stub(:find_by_permalink!)
-      get :show, { :id => 'about' }
-      assigns(:admin).should be_true
+    it 'assigns a persisted page as @page' do
+      expect(Page).to receive(:find_by_permalink!).with('about') { about_page }
+      get :show, {id: 'about'}
+      expect(assigns(:page)).to eq about_page
     end
 
-    it 'assigns @page if params[id] is a valid permalink' do
-      Page.should_receive(:find_by_permalink!).with('about').and_return(@valid_page)
-      get :show, { :id => 'about'}
-      assigns(:page).should eq @valid_page
+    before do
+      allow(controller).to receive(:current_user)
+      allow(Page).to receive(:find_by_permalink!)
     end
 
-    it 'assigns 404 page if params[id] is not a valid permalink' do
-      Page.should_receive(:find_by_permalink!).with('lalala').and_raise(ActiveRecord::RecordNotFound)
-      Page.should_receive(:find_by_permalink!).with('404').and_return(@error_page)
-      get :show, { :id => 'lalala'}
-      assigns(:page).should eq @error_page
+    describe 'assigns site superadmin status of current_user to @superadmin' do
+      it 'NIL when there is no current_user' do
+        get :show, {id: 'about'}
+        expect(assigns(:superadmin)).to be nil
+      end
+      it 'FALSE when current_user is NOT superadmin' do
+        allow(user).to receive(:superadmin?) { false }
+        allow(controller).to receive(:current_user) { user }
+        get :show, {id: 'about'}
+        expect(assigns(:superadmin)).to be false
+      end
+
+      it 'TRUE when current_user is superadmin' do
+        allow(user).to receive(:superadmin?) { true }
+        allow(controller).to receive(:current_user) { user }
+        get :show, {id: 'about'}
+        expect(assigns(:superadmin)).to be true
+      end
     end
 
-    it '200: respond code should be appropriate for page' do
-      Page.stub(:find_by_permalink!)
-      get :show, { :id => 'about' }
-      response.status.should eq 200
+    it 'assigns a persisted page as @page' do
+      expect(Page).to receive(:find_by_permalink!).with('about') { about_page }
+      get :show, {id: 'about'}
+      expect(assigns(:page)).to eq about_page
     end
 
-    it '404: respond code should be appropriate for page' do
-      Page.should_receive(:find_by_permalink!).with('lalala').and_raise(ActiveRecord::RecordNotFound)
-      Page.should_receive(:find_by_permalink!).with('404').and_return(@error_page)
-      get :show, { :id => 'lalala'}
-      response.status.should eq 404
+    it 'is NOT restricted' do
+      expect(controller).not_to receive(:superadmin?)
+      get :show, {id: 'about'}
+      expect(response.status).to eq 200
+    end
+
+    it 'uses a full-width layout' do
+      get :show, {id: 'about'}
+      expect(response).to render_template 'layouts/full_width'
     end
   end
 
-  describe "GET new" do
-    it "assigns a new page as @page" do
+  describe 'GET new' do
+    before { allow(Page).to receive(:new) }
+
+    it 'assigns a new page as @page' do
+      expect(Page).to receive(:new) { page }
       get :new, {}
-      assigns(:page).should be_a_new(Page)
+      expect(assigns(:page)).to eq page
     end
 
-    it 'should use a full-width layout' do
+    it 'is restricted' do
+      expect(controller).to receive(:superadmin?) { false }
       get :new, {}
-      response.should render_template 'layouts/full_width'
+      expect(response.status).to eq 302
+    end
+
+    it 'uses a full-width layout' do
+      get :new, {}
+      expect(response).to render_template 'layouts/full_width'
     end
   end
 
-  describe "GET edit" do
-    it "assigns the requested page as @page" do
-      page = Page.create! valid_attributes
-      get :edit, {:id => page.to_param}
-      assigns(:page).should eq(page)
+  describe 'GET edit' do
+    before { allow(Page).to receive(:find_by_permalink!) }
+
+    it 'assigns the requested page as @page' do
+      expect(Page).to receive(:find_by_permalink!).with(page.id) { page }
+      get :edit, {id: page.id}
+      expect(assigns(:page)).to eq page
     end
 
-    it 'should use a full-width layout' do
-      page = Page.create! valid_attributes
-      get :edit, {:id => page.to_param}
-      response.should render_template 'layouts/full_width'
-    end
-  end
-
-  describe "POST create" do
-    describe "with valid params" do
-      it "creates a new Page" do
-        expect {
-          post :create, {:page => valid_attributes}
-        }.to change(Page, :count).by(1)
-      end
-
-      it "assigns a newly created page as @page" do
-        post :create, {:page => valid_attributes}
-        assigns(:page).should be_a(Page)
-        assigns(:page).should be_persisted
-      end
-
-      it "redirects to the created page" do
-        post :create, {:page => valid_attributes}
-        response.should redirect_to(Page.last)
-      end
+    it 'is restricted' do
+      expect(controller).to receive(:superadmin?) { false }
+      get :edit, {id: page.id}
+      expect(response.status).to eq 302
     end
 
-    describe "with invalid params" do
-      it "assigns a newly created but unsaved page as @page" do
-        # Trigger the behavior that occurs when invalid params are submitted
-        Page.any_instance.stub(:save).and_return(false)
-        post :create, {:page => { "name" => "invalid value" }}
-        assigns(:page).should be_a_new(Page)
-      end
-
-      it "re-renders the 'new' template" do
-        # Trigger the behavior that occurs when invalid params are submitted
-        Page.any_instance.stub(:save).and_return(false)
-        post :create, {:page => { "name" => "invalid value" }}
-        response.should render_template("new")
-      end
+    it 'uses a full-width layout' do
+      get :edit, {id: page.id}
+      expect(response).to render_template 'layouts/full_width'
     end
   end
 
-  describe "PUT update" do
-    describe "with valid params" do
-      it "updates the requested page" do
-        page = Page.create! valid_attributes
-        # Assuming there are no other pages in the database, this
-        # specifies that the Page created on the previous line
-        # receives the :update_attributes message with whatever params are
-        # submitted in the request.
-        Page.any_instance.should_receive(:update_attributes).with({ "name" => "MyString" })
-        put :update, {:id => page.to_param, :page => { "name" => "MyString" }}
-      end
+  describe 'POST create' do
+    let(:attributes) { {name: 'Yoohoo!', permalink: 'yoohoo', content: 'yada yada'} }
 
-      it "assigns the requested page as @page" do
-        page = Page.create! valid_attributes
-        put :update, {:id => page.to_param, :page => valid_attributes}
-        assigns(:page).should eq(page)
-      end
-
-      it "redirects to the page" do
-        page = Page.create! valid_attributes
-        put :update, {:id => page.to_param, :page => valid_attributes}
-        response.should redirect_to(page)
-      end
+    before do
+      allow(Page).to receive(:new) { page }
+      allow(page).to receive(:save)
     end
 
-    describe "with invalid params" do
-      it "assigns the page as @page" do
-        page = Page.create! valid_attributes
-        # Trigger the behavior that occurs when invalid params are submitted
-        Page.any_instance.stub(:save).and_return(false)
-        put :update, {:id => page.to_param, :page => { "name" => "invalid value" }}
-        assigns(:page).should eq(page)
-      end
+    it 'assigns a newly created page to @page' do
+      expect(Page).to receive(:new).with(attributes.stringify_keys) { page }
+      post :create, {page: attributes}
+      expect(assigns(:page)).to eq page
+    end
 
-      it "re-renders the 'edit' template" do
-        page = Page.create! valid_attributes
-        # Trigger the behavior that occurs when invalid params are submitted
-        Page.any_instance.stub(:save).and_return(false)
-        put :update, {:id => page.to_param, :page => { "name" => "invalid value" }}
-        response.should render_template("edit")
-      end
+    it 'if created page is valid, it redirects to show it' do
+      expect(page).to receive(:save) { true }
+      post :create, {page: attributes}
+      expect(response).to redirect_to page
+      expect(flash[:notice]).to eq 'Page was successfully created.'
+    end
+
+    it 'if created page is INVALID, it re-renders the "new" template' do
+      expect(page).to receive(:save) { false }
+      post :create, {page: attributes}
+      expect(response).to render_template 'new'
+    end
+
+    it 'is restricted' do
+      expect(controller).to receive(:superadmin?) { false }
+      post :create, {page: attributes}
+      expect(response.status).to eq 302
+    end
+
+    it 'uses a full-width layout' do
+      post :create, {page: attributes}
+      expect(response).to render_template 'layouts/full_width'
     end
   end
 
-  describe "DELETE destroy" do
-    it "destroys the requested page" do
-      page = Page.create! valid_attributes
-      expect {
-        delete :destroy, {:id => page.to_param}
-      }.to change(Page, :count).by(-1)
+  describe 'PUT update' do
+    let(:attributes) { {name: 'Yoohoo!', permalink: 'yoohoo', content: 'yada yada'} }
+
+    before do
+      allow(Page).to receive(:find_by_permalink!) { page }
+      allow(page).to receive(:update_attributes)
     end
 
-    it "redirects to the pages list" do
-      page = Page.create! valid_attributes
-      delete :destroy, {:id => page.to_param}
-      response.should redirect_to(pages_url)
+    it 'assigns the updated page to @page' do
+      expect(Page).to receive(:find_by_permalink!).with(page.id) { page }
+      put :update, {id: page.id, page: attributes}
+      expect(assigns(:page)).to eq page
+    end
+
+    it 'if updated page is valid, it redirects to show it' do
+      expect(page).to receive(:update_attributes).with(attributes.stringify_keys) { true }
+      put :update, {id: page.id, page: attributes}
+      expect(response).to redirect_to page
+      expect(flash[:notice]).to eq 'Page was successfully updated.'
+    end
+
+    it 'if updated page is INVALID, it re-renders the "edit" template' do
+      expect(page).to receive(:update_attributes).with(attributes.stringify_keys) { false }
+      put :update, {id: page.id, page: attributes}
+      expect(response).to render_template 'edit'
+    end
+
+    it 'is restricted' do
+      expect(controller).to receive(:superadmin?) { false }
+      put :update, {id: page.id, page: attributes}
+      expect(response.status).to eq 302
+    end
+
+    it 'uses a full-width layout' do
+      put :update, {id: page.id, page: attributes}
+      expect(response).to render_template 'layouts/full_width'
     end
   end
+  
+  describe 'DELETE destroy' do
+    before do
+      allow(Page).to receive(:find_by_permalink!) { page }
+      allow(page).to receive(:destroy)
+    end
 
+    it 'assigns the page to be destroyed to @page' do
+      expect(Page).to receive(:find_by_permalink!).with(page.id) { page }
+      delete :destroy, {id: page.id}
+      expect(assigns(:page)).to eq page
+    end
+
+    it 'destroys the page' do
+      expect(page).to receive(:destroy)
+      delete :destroy, {id: page.id}
+    end
+
+    it 'redirects to the index of pages' do
+      delete :destroy, {id: page.id}
+      expect(response).to redirect_to pages_path
+    end
+
+    it 'is restricted' do
+      expect(controller).to receive(:superadmin?) { false }
+      delete :destroy, {id: page.id}
+      expect(response.status).to eq 302
+    end
+  end
+  describe ".permit" do
+    it "returns the cleaned params" do
+      pages_params = { page: {content: 'stuff', name: 'this', permalink: 'here', link_visible: false}}
+      params = ActionController::Parameters.new.merge(pages_params)
+      permitted_params = PagesController::PageParams.build(params)
+      expect(permitted_params).to eq(pages_params.with_indifferent_access[:page])
+    end
+  end
 end
