@@ -13,6 +13,7 @@ module ApplicationHelper
     red_carpet = Redcarpet::Markdown.new(Redcarpet::Render::HTML, :autolink => true)
     red_carpet.render(text).html_safe
   end
+
   def resource_name
     :user
   end
@@ -25,17 +26,31 @@ module ApplicationHelper
     @devise_mapping ||= Devise.mappings[:user]
   end
 
-  def after_sign_in_path_for(resource)
-    redirect_to  root_url
-  end
-
-  def parent_layout(layout)
+  def parent_layout(layout) # http://andre.arko.net/2013/02/02/nested-layouts-on-rails--31/
     @view_flow.set(:layout, output_buffer)
     self.output_buffer = render(:file => "layouts/#{layout}")
   end
 
   def cookie_policy_accepted?
     cookies['cookie_policy_accepted'].present?
+  end
+
+  def active_if(controller)
+    'active' if params[:controller] == controller
+  end
+
+  def feature_active?(flag)
+    Feature.active?(flag.to_sym)
+  end
+  def bootstrap_class_for flash_type
+    case flash_type
+    when "warning"
+     "alert-warning"
+    when "notice"
+     "alert-success"
+    else
+      "alert-error"
+    end
   end
 end
 
