@@ -8,14 +8,11 @@ LocalSupport::Application.configure do
   config.cache_classes = true
 
   # Configure static asset server for tests with Cache-Control for performance
-  config.serve_static_assets = true
+  config.serve_static_files = true
   config.static_cache_control = "public, max-age=3600"
 
-  # Log error messages when you accidentally call methods on nil
-  config.whiny_nils = true
-
   # Show full error reports and disable caching
-  config.consider_all_requests_local       =  false
+  config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
 
   # Raise exceptions instead of rendering exception templates
@@ -40,6 +37,8 @@ LocalSupport::Application.configure do
   # Allow pass debug_assets=true as a query parameter to load pages with unpackaged assets
   config.assets.allow_debugging = true
 
-  config.action_mailer.default_url_options = { :host => 'localhost:3000' } 
+  config.action_mailer.default_url_options = { :host => 'localhost:3000', :port => 3000 }
+  Rails.application.routes.default_url_options[:host] = 'localhost:3000'
   config.middleware.use RackSessionAccess::Middleware
+  config.eager_load = false
 end
